refactor: address reconciliation review feedback
